requirements.txt
gradio
icrawler
facenet-pytorch
lxml
beautifulsoup4
ftfy
regex
tqdm
repackage
git+https://github.com/openai/CLIP.git
